backend/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
python-dotenv
redis
websockets
pytest
httpx
python-jose
passlib
python-multipart
autogen
